Fixes for Visual Studio

+++ native/src/seal/util/fips202.c
@@ -67,7 +67,8 @@ static void store64(uint8_t x[8], uint64_t u) {
   unsigned int i;
 
   for(i=0;i<8;i++)
-    x[i] = u >> 8*i;
+    /* Microsoft SEAL edit: explicit cast to silence warnings. */
+    x[i] = (uint8_t)(u >> 8*i);
 }
 
 /* Keccak round constants */
